+++ src/Symfony/Component/HttpFoundation/CHANGELOG.md
@@ -18,6 +18,7 @@ CHANGELOG
  * Make `HeaderBag::getDate()`, `Response::getDate()`, `getExpires()` and `getLastModified()` return a `DateTimeImmutable`
  * Support root-level `Generator` in `StreamedJsonResponse`
  * Add `UriSigner` from the HttpKernel component
+ * Add `partitioned` flag to `Cookie` (CHIPS Cookie)
 
 6.3
 ---